fix: move template calculation to PaymentTable so it is shown on Payment and PaymentCalcTable
