+++ scripts/src/form/confirmation/PaymentTable.d.ts
@@ -1,5 +1,9 @@
 import { IPaymentInfo } from "../interfaces";
 
+export interface IFormData {
+  [x: string]: any;
+}
 export interface IPaymentTableProps {
   paymentInfo: IPaymentInfo;
+  formData: IFormData;
 }
