+++ scripts/src/form/confirmation/payment.tsx
@@ -5,7 +5,6 @@ import CCAvenue from "./CCAvenue";
 import ManualApproval from "./ManualApproval";
 import sanitize from "../../sanitize";
 import ExpressionParser from "../../common/ExpressionParser";
-import nunjucks from "nunjucks";
 import { IPaymentProps, IPaymentInfo } from "../interfaces";
 
 let Components = {
@@ -13,20 +12,6 @@ let Components = {
   ccavenue: CCAvenue,
   manual_approval: ManualApproval,
   manual_approval_2: ManualApproval
-};
-
-nunjucks.installJinjaCompat();
-
-const processCurrency = (paymentInfo, formData) => {
-  if (paymentInfo.currencyTemplate) {
-    let currency = sanitize(
-      nunjucks.renderString(paymentInfo.currencyTemplate, {
-        value: formData || {}
-      })
-    );
-    return { ...paymentInfo, currency };
-  }
-  return paymentInfo;
 };
 
 class Payment extends React.Component<IPaymentProps, any> {
@@ -118,10 +103,8 @@ class Payment extends React.Component<IPaymentProps, any> {
             this.props.paymentInfo.items &&
             this.props.paymentInfo.items.length > 0 && (
               <PaymentTable
-                paymentInfo={processCurrency(
-                  this.props.paymentInfo,
-                  this.props.formData
-                )}
+                paymentInfo={this.props.paymentInfo}
+                formData={this.props.formData}
               />
             )}
           {this.props.paymentInfo_received &&
